fix(playwright): No DOMParser available
For Playwright tests, we don't have jsdom at hand, and thus, we
cannot create a DOMParser instance.

Moved corresponding code to extra file `RichTextDOM` which can
now be used by tests or for example-data creation where JSDOM
is available.
