+++ packages/ckeditor5-coremedia-example-data/src/RichTextBase.ts
@@ -446,15 +446,3 @@ export const richtext = (
   result += `>${innerXml}</div>`;
   return result;
 };
-
-const domParser = new DOMParser();
-
-export const parseRichTextFromString = (
-  innerXml: Content = "",
-  enforcedNamespaces: ReadonlyArray<string> = []
-): Document => {
-  const xml = richtext(innerXml, true, enforcedNamespaces);
-  return domParser.parseFromString(xml, "text/xml");
-};
-
-export const richTextDocument = parseRichTextFromString("", ["xlink"]);
