+++ packages/ckeditor5-coremedia-example-data/src/RichTextDOM.ts
@@ -1,0 +1,13 @@
+import { Content, richtext } from "./RichTextBase";
+
+const domParser = new DOMParser();
+
+export const parseRichTextFromString = (
+  innerXml: Content = "",
+  enforcedNamespaces: ReadonlyArray<string> = []
+): Document => {
+  const xml = richtext(innerXml, true, enforcedNamespaces);
+  return domParser.parseFromString(xml, "text/xml");
+};
+
+export const richTextDocument = parseRichTextFromString("", ["xlink"]);
